+++ Calculadora/Origem.cpp
@@ -2,36 +2,36 @@
 //subtrair, dividir e multiplicar
 #include <iostream>
 
-long double Sum(int a, int b);
-long double Mult(int a, int b);
-long double Div(int a, int b);
-long double Subt(int a, int b);
+long double Sum(float a, float b);
+long double Mult(float a, float b);
+long double Div(float a, float b);
+long double Subt(float a, float b);
 
 
-long double Sum(int a, int b)
+long double Sum(float a, float b)
 {
-	int FinalResult = 0;
+	double FinalResult = 0;
 	FinalResult += a + b; //adiciona ao Resultado anterior o valor da nova operação
 	return FinalResult;
 }
 
-long double Mult(int a, int b)
+long double Mult(float a, float b)
 {
-	int FinalResult = 0;
+	double FinalResult = 0;
 	FinalResult += (a) * (b); //adiciona ao Resultado anterior o valor da nova operação
 	return FinalResult;
 }
 
-long double Div(int a, int b)
+long double Div(float a, float b)
 {
-	int FinalResult = 0;
+	double FinalResult = 0;
 	FinalResult += (a) / (b);//adiciona ao Resultado anterior o valor da nova operação
 	return FinalResult;	
 }
 
-long double Subt(int a, int b)
+long double Subt(float a, float b)
 {
-	int FinalResult = 0;
+	double FinalResult = 0;
 	FinalResult += (a) - (b);//adiciona ao Resultado anterior o valor da nova operação
 	return FinalResult;
 }
